for the Monte-Carlo simulation with sub-systems we need something
like a simulation hierarchy: a tree structure of the matrix data
of the host-system and the sub-systems. In the simulation wee
first simulate and calculate the matrix data of the leafs of this
tree structure than integrate the result into the matrix data of
the next level, simulate and calculate them etc. until this is
done for the final host-system.

We can use a shared matrix with all uncertainty distributions of
the host-system and all sub-systems as these distributions of
input, output, and process parameter amounts are indipendent
from how the product system is linked. However, we have to make
sure the the parameter redefinitions are applied correctly.

We also assume that this matrix with uncertainty distributions
is sparse or small.

... but before we can go into this, we need to make the API more
powerful