creating an `CalcExchange` should go into a class like
`ExchangeTable` as we want to get rid of the exchange cache.
I mean the cache is nice because we do not have to touch the
database for repeated calculations, but this does not scale
memory-wise with huge dense matrices like in PSILCA.

also, it is not very readable but accessing the records in an SQL
result by index is much faster than using the field names...

And should we support formulas for distribution parameters? We have
the fields for this in the model but this could really be confusing:
a distribution parameter could be a formula with a reference to an
input parameter which again has an uncertainty distribution with
parameters that can have formulas etc. ... I will not implement this
for now as there is probably no use case. We could even think about
removing these fields from the model in future versions.

when we access the SQL fields by index is it helpful to have the
index field assignment in the code, we can generate that:

```dart
void main() {
  var fields = [
    "id",
    "f_owner",
    "f_flow",
    "f_flow_property_factor",
    "f_unit",
    "resulting_amount_value",
    "resulting_amount_formula",
    "is_input",
    "avoided_product",
    "f_default_provider",
    "cost_value",
    "cost_formula",
    "f_currency",
    "distribution_type",
    "parameter1_value",
    "parameter2_value",
    "parameter3_value"
  ];
  var query = '"SELECT" \n';
  for (int i = 0; i < fields.length; i++) {
    query += '+ /* $i */ " ${fields[i]}"\n';
  }
  query += '+ " FROM tbl_exchanges";';
  print(query);
}
```

finally, we need to do some benchmarks of the ExchangeTable with real
databases (when we replaced the FlowTable with a simple map)