ok, move the building process of the `FlowIndex` into the inventory
builder. woah ... we already know in the inventory builder what should
go into the intervention matrix ... thus we do not have to repeat all
the checks in the `FlowIndexBuilder`. The `FlowIndexBuilder` can be
removed when we take the new inventory builder -> tagged it as
`@Deprecated`. To build the flow index we need the descriptors -> we
really need some faster methods to load them from the database...