the formula interpreter should be created outside of the inventory
builder because it could contain all kind of parameter redefinitions
from product systems, projects etc. With this, our configuration
object already looks like this:

```java
public class InventoryConfig {
  public IDatabase db;
  public TechIndex index;
  public boolean withUncertainties;
  public AllocationMethod allocationMethod;
  public Map<ProcessProduct, SimpleResult> subResults;
  public FormulaInterpreter interpreter;
}
```

it is a pity that Java does not have named parameters; we have to
introduce a class for this ...

on the other side, I already prepared the `ParameterTable` to be
ready for Monte-Carlo simulations. So passing the interpreter as
parameter into the inventory builder is a good idea.

but before we can evaluate the formulas we need to find a way to
store them for the Monte-Carlo simulations...