there are quite some cases that we have to consider when we need
to merge uncertainty cells -> e.g. the same elementary flow on the
input and output side with different distributions ... but we will
handle them when we are ready with the base case.

an `UCell` can simply wrap an `CalcExchange` -> it can use the new
`matrixValue` value method after generating a new amount value;
the formula interpreter with the new generated parameter
distributions wil be passed into each iteration step;

ahh ... for multiple distributions that map to the same matrix cell,
we could have a list of UCells ... attached to a host-UCell ... the
final matrix value is then just the sum off all the values generated
by the respective UCells -> this could work.

... the UCell needs to hold the allocation factor; when there are
multiple UCells attached to a host cell, each cell could have a
different allocation factor (in case of causal allocation). So these
would be the next trivial steps:

* implement the `matrixValue` and `costValue` methods in `CalcExchange`
  (the conversion table may should also contain conversion factors
   of currencies and we may need to add a conversion factor for cost
   values to the `CalcExchange`?)
* in the inventory builder we can replace the `mergeExchange` method
  by just adding up the matrix values -> super simple.
* `UCell` wraps a `CalcExchange` and could have an attached list of
  other UCells; with an allocation factor for each cell -> when there
  are multiple distributions mapped to a single matrix cell we just
  add the generated values up after a simulation run.
* In a simulation step we also pass the formula interpreter into
  the UCells ...