we should move the conversion factors for currencies also into the
`CalcExchange` and collect it when we go through the `ExchangeTable`.
First step would be to merge the `ConversionTable` with the `CurrencyTable`.
We will remove the currencyID from the `CalcExchange` and replace it with
a `currencyFactor`. In the `InventoryConfig` we add a flag `withCosts` which
decides whether to build the cost vectors in the `MatrixData`.