mhh ... we have to consider uncertain parameters that are used in formulas
for costs ... -> we pass the `InventoryConfig` into the `UMatrix` to get
the information if we should handle this case.

ahh ... we could have the case where multiple exchanges are mapped to the
same matrix cell and not all exchanges have an uncertainty distribution (or
formula assigned). In this case we cannot decide whether an exchange should
be added to the `UMatrix` when building the matrix but only when we already
collected all exchanges. We could have a something like a `compact` method
which removes exchanges with no uncertainty (no distribution and no formulas
(also for costs)) from the `UMatrix` after the build (-> we need to check
whether this has some advantages)