Currently we have to pass around the `MatrixCache` everywhere because
it is used in `DataStructures` etc... we will get rid of this in
future versions.

This is now the algorithm for initializing the simulation graph. We
have a topological order of simulation nodes of the sub-systems.
The sub-system nodes do not have LCIA matrices attached as they are
just used for LCI (and LCC) calculation.

this will need quite some tests ... but the concept is beautiful.

next we have to add the `UMatrix` instances to `MatrixData` and
the dispatch the simulation calls down to the `UCell`s