ok, back to MC simulation: first we need to build the `UMatrix`
instances in the new inventory builder. These are then attached
to the `MatrixData`. The call tree in the MC simulation would
then be:

```
Type                      Method
-> SimulationGraph        nextRun()
  -> SimulationNode       nextRun()
    -> MatrixData         simulate(FormulaInterpreter)
      -> UMatrix          generate(IMatrix, FormulaInterpreter)
        -> UCell          generate(FormulaInterpreter)
```

In each run, a simulation node needs to fill up the matrix
columns that are linked to sub-systems -> because of the
topological ordered nodes it is assured that the results
for this are already calculated.

To get this done, we may do first the number generation of
the LCIA factors in the `SimulationGraph` but later we
should move the current `ImpactFactorMatrix` etc. also to
the new `MatrixBuilder` based structure.