Lets test the `SimulationGraph` of a simple system
we have a system `A` that has a single process `A` with an
output `a` and a system `B` also with a single process `B`
and output `b`. Process `B` has an input of `a` with a
distribution of `uniform(1, 2)` which is linked in the
product system `B` to the sub-system `A`. Process `A` has
elementary flow emissions of `uniform(3, 6)` related to
one unit of `a`.

```
+---+    +---+
| A | -> | B |
+---+    +---+
```

When running the simulation we should get values between
[3, 12] for `e`...

... no, the current implementation will not work. We need
to pass the indices (`FlowIndex`, `TechIndex`) via the
results into the `InventoryBuilder`. The result can be
empty but we need the correct indices ... we need to build
the matrix data in topological order ... and the problem
is that we are currently use the matrix data to identify
the sub-system links -> need to change that.