I am still not sure whether we need the `isSystemLink` flag in
`ProcessLink`. This is redundant information as we know the ID of
the provider and can figure out the type of the provider (process
or product system). It only makes sense to use this field if this
makes something faster or much easier; but here I am not sure.
On the other side it is just too easy to have the wrong
information in this field.

Regarding the performance, we could run a benchmark to check
whether the flag helps in code like this:

```java
// allRels contains the sub-system relations of each product system
// in the database as: hostSystemID -> (subSystemID, hostSystemID)*
Map<Long, List<LongPair>> allRels = new HashMap<>();
sql = "select f_product_system, f_provider from tbl_process_links";
NativeSql.on(db).query(sql, r -> {
  long provider = r.getLong(2);
  if (!systems.contains(provider))
    return true;
  long system = r.getLong(1);
  List<LongPair> rels = allRels.get(system);
  if (rels == null) {
    rels = new ArrayList<>();
    allRels.put(system, rels);
  }
  rels.add(LongPair.of(provider, system));
  return true;
});
```

The initialization code reminds me of this:
https://news.ycombinator.com/item?id=18772873

... with "pinned process contributions" we do not need to
different iteration results via the API, do we?