## Calculating the contributions of pinned providers
We want to calculate the direct and upstream contributions of
pinned providers in a Monte-Carlo simulation. Pinned providers
means processes and sub-systems of the product system for which
the MC simulation is executed. In each simulation run we get
the scaling vector $s$. We also have the technology and
intervention matrix always in the state of the last run.

For a pinned provider $i$ the direct LCI contributions $d_i$
can be calculated by scaling the column $i$ of the intervention
matrix with the respective scaling factor $s_i$:

$$d_i = s_i \odot B[:,i]$$

For the upstream contribution of $i$ we first need the total
requirements of $i$:

$$t_i = s_i \cdot A[i,i]$$

With this, we calculate the upstream contributions $u_i$ as:

$$s^u_i = \text{solve}(A,f^t_i) \ , u_i = B s^u_i$$

In this equation, $f^t_i$ is just the demand vectors that
contains a single non-zero entry with the value $t_i$ at
position $i$.

To get the LCIA contributions, we just need to multiply the
matrix with the characterization factors $F$ with the
respective LCI contributions. So, $F \ d_i$ gives the
direct and $F \ u_i$ the upstream LCIA contributions.

We could store these contributions of the different
simulation runs as vector of lists:

```java
class PinnedContributions {
  ProcessProduct product;
  List<double[]> directLCI;
  List<double[]> upstreamLCI;
  List<double[]> directLCIA;
  List<double[]> upstreamLCIA;
}
```

In the `SimulationResult` the results are currently stored
as arrays of `List<Double>` but I think a `List<double[]>`
is better.

Also, some methods like these would be helpful:

```java
double get(FlowDescriptor d, int i);
double get(ImpactCategoryDescriptor d, int i);
```

ok, let me refactor this first