the `SimulationStatistics` currently mixes in the things that are
required for the display in the chart. Better would be something
like `int[] getFrequencies(double[] values, int intervalCount)`.

it's tiring to write all these accessor methods ... I should
really review and test this later... and there are calculation
errors with waste flows in the tests.

the calculation of the pinned contributions above needs to
include the **loop factor** of the reference product. We need
to test this with systems without uncertainty distributions to
verify that the basic calculation is correct.