forgot to add the contributions to the simulation result. now
it works. A minimum size for the simulation chart would be nice.

Also, we have to clean-up the Julia solver:

* write a general CCR implementation of the IMatrix interface
  that can be directly used in UMFPACK and friends
* write bindings for sparse matrix-matrix multiplications
* we need to get the Julia bindings running on all platforms

-> would be cool if we could use Rust for writing these
bindings:

* with `cargo` we could manage the build process on all
  platforms
* there is a JNI package for Rust which makes it easier to
  write JNI bindings
* the Rust LSP implementation gives cross-platform editor
  support

-> I will test this in an experimental project.

the UMFPACK libraries are GPL licensed. we cannot distribute
them with openLCA as we are running on the Eclipse RCP and
the GPL and EPL are incompatible:
https://eclipse-foundation.blog/2010/04/06/epl-gpl-commentary/

-> we need to think about how we can make it as easy as
possible for the users to add these libraries to openLCA on
their own.