complexity comes with the updates: when we want to update things
that have already an ID that is used in other entities we need
to make sure that such IDs remain stable. However, it is possible
to break contracts: the flow of an exchange that is linked by
a product system as reference flow could get a new flow property
with the old flow property removed.

in the app, we could add the source jars for debugging:

```xml
<execution>
  <id>src-dependencies</id>
  <phase>package</phase>
  <goals>
    <goal>copy-dependencies</goal>
  </goals>
  <configuration>
    <classifier>sources</classifier>
    <failOnMissingClassifierArtifact>false</failOnMissingClassifierArtifact>
    <outputDirectory>${basedir}/libsrc</outputDirectory>
  </configuration>
</execution>
```