the error that an update of an LCIA method over IPC removes all
of its LCIA categories should be fixed now. However, an update of
an LCIA method over IPC does not update the LCIA categories (e.g.
when new LCIA factors were added) as the LCIA categories are not
attached to the LCIA method in the JSON-LD format. From the
design of the format, this is good as we want to move to
stand-alone LCIA categories in the future. However, currently
LCIA categories are always attached to a single LCIA method in
openLCA. This is the dilemma regarding the IPC update: we need
to update the LCIA method when we want to update the LCIA
category & factors but the LCIA method does not contain the
JSON-LD data for the update.

To solve this, I added the update functionality for LCIA
categories over the IPC API which then updates the respective
method. Here is an example that I used to test this:

```python
import olca

client = olca.Client(8080)

flow = client.get(olca.Flow,
                  'dfe445cf-568d-4623-b344-865feb07149e')

# find and update the LCIA category
method = client.get(olca.ImpactMethod,
                    'd5a299bb-61e9-49f8-b4b5-761b95e8ce59')
impact = None
for icat in method.impact_categories:
    if icat.name == 'GWP':
        impact = client.get(olca.ImpactCategory, icat.id)
        break

# add the new LCIA factor
impact_factor = olca.ImpactFactor()
impact_factor.flow = flow
impact_factor.value = 24.0
if impact.impact_factors is None:
    impact.impact_factors = []
impact.impact_factors.append(impact_factor)
client.update(impact)
```

There are probably more things that need to be fixed in the
IPC API. It is based on the standard JSON-LD import/export
functions but these functions expect that there is always
an entity store present which contains all referenced data
which is not the case when going the IPC way.