we can make the creation of categories in the
imports more simple: in general we only need
to sync a category path with the database.
we can put this function directly into the
`CategoryDao` and use it in the various imports.

in the imports, we then just need to extract the
respective path as string array...

```clojure
(-> (java.util.Date.) (.getTime))
```