the `CalcParameter` is never used? Also, let us remove all
these getters and setters from the `CalcImpactFactor`.

We will move the LCIA factor matrices to a system that is
equivalent to the new exchange matrices... but first we
have to make LCIA categories a first class citizen (means
a stand-alone entity). With this the design will be much
cleaner, e.g. the formula scope is the one of the LCIA
category which is the owner of the LCIA factor like
it is the process for an exchange. The `CalcImpactFactor`
can then have an equivalen method
`getMatrixValue(FormulaInterpreter)` like the `CalcExchange`

the main problem for this redesign are the regionalized
methods: shape files and parameters are bound to an LCIA
method, not to the LCIA categories. it would require some
work to fix this but the result should be much better.