can we replace the location ID in the flow and process
descriptors directly with the location code (as string)?
(later also the other IDs: like for the category and
flow property?)

we should really benchmark this...

in FlowDao.getDescriptors() we could first create a category
map so that flows in the same category just have a pointer
to the same path: this should not cost much additional memory
but with this we could throw out the caches?