we should not throw a warning for each missing flow
in the ILCD LCIA method import. There are just to much of those errors
in Gabi packages... we will just throw these things when the log level
is set to `trace`. Also, when in the characterization factors of mapped
flows the conversion factors were not applied... hope I correctly fixed
that.