we should save the flow states into the mapping file
the mapping tool should be then should check whether these
states are still valid => this is some advanced workflow that
I will implement later today.

for saving the states, we just need two additional columns in
the mapping file; later also in the JSON-LD format.

=> 20 is currently the last index

```java
row[20] = t.providerLocation;
```

so we put the status of the source flow into `[21]` and the
status of the target flow into `[22]`