moving the database upgrades back to the core
this is currently not used in the application but we will move the
database
upgrade functions back to the core module. the idea is to remove the
current
database update manager that uses a specific package format where a
Python
script is executed to run an update. the initial idea of this was that
we
can not only update the database schema but also the data in a database
with
attached data files in such an update package. however, we used this
data
update functionality and now there is the collaboration server that
exactly
provides this functionality.