in the last update we added some columns that are not
used yet; mainly regarding LCIA methods. however, we
want to make bigger changes to the LCIA model in openLCA
in the future (LCIA categories as stand-alone entities
that can be shared between LCIA methods). also, I am not
sure if each column in each column is really used (e.g.
f_currency in the process table) => we should check this
in the next schema update.