example from the UMFPACK manual:
https://fossies.org/linux/SuiteSparse/UMFPACK/Doc/UMFPACK_UserGuide.pdf

```c++
int n = 5 ;
int Ap [ ] = {0, 2, 5, 9, 10, 12} ;
int Ai [ ] = { 0, 1, 0, 2, 4, 1, 2, 3, 4, 2, 1, 4} ;
double Ax [ ] = {2., 3., 3., -1., 4., 4., -3., 1., 2., 2., 6., 1.} ;
double b [ ] = {8., 45., -3., 3., 19.} ;
double x [5] ;
```

so the last entry in the column pointers is the number of non-zero values
$nz(A)$ and not $nz(A) + 1$

```
The column pointers of the column-oriented form of the matrix A. See
umfpack_*_symbolic for a description. The number of entries in
the matrix is nz = Ap [n_col]. Ap [0] must be zero, Ap [n_col] must be
=> 0, and Ap [j] <= Ap [j+1] and Ap [j] <= Ap [n_col] must be true for
all j in the range 0 to n_col-1. Empty columns are OK (that is, Ap [j]
may equal Ap [j+1] for any j in the range 0 to n_col-1).
```

but here it is $nz(A) + 1$ again:

https://eclass.uoa.gr/modules/document/file.php/D34/%CE%A4%CE%95%CE%A3%CE%A4_%CE%A0%CE%99%CE%9D%CE%91%CE%9A%CE%95%CE%A3/p1-duff.pdf

but in SciPy it also seems to be the plain number of non-zero entries