offset calculation in C-order
we always write our matrices in Fortran order so this is
not tested. also, it requires $A.rows$ disk seeks to read
a column from a matrix file stored in C-order.

```groovy
int rows = 2
int cols = 3
// data stored in C-order
double[] data = [1.0, 2.0, 3.0, 4.0, 5.0, 6.0]

int col = 1

for (row in 0..rows-1) {
  int offset = row * cols + col
  println data[offset]
}
```

similar to `repl` there is a `wfrcl` when working with buffers:

read
  -> eval
    -> print
      -> loop

write (the buffer)
  -> flip (the buffer)
    -> read (the buffer)
      -> clear (the buffer)
        -> loop