prepare `UMatrix` so that it can also take LCIA factors
while exchanges (also with different uncertainty distributions) can be
mapped to the same matrix cell this is not true of LCIA factors. A LCIA
category with two different factors for the same flow is not valid. we
will not throw or log an error here because this is something that
should be catched in the user interface or import. we just take the last
value in such a case.