getting a single openLCA LCIA method for an ILCD LCIA indicator
is not valid anymore. thus, we need to find a better way to map
LCIA indicators to methods in the import. unfortunately, LCIA
method names are stored as plain strings in the ILCD datasets;
no UUIDs can be assigned.

```java
public ImpactMethod getFromExtension() {
   QName refIdAtt = new QName("http://openlca.org/ilcd-extensions",
        "olca_method_uuid");
  String refId = ilcdMethod.otherAttributes.get(refIdAtt);
  if (refId == null)
    return null;
  ImpactMethod method = dao.getForRefId(refId);
  return method != null ? method : createFromExtension(refId);
}
```

also, the type is better added to the LCIA indicator?

```java
private String getType() {
  Modelling mav = ilcdMethod.modelling;
  if (mav == null || mav.normalisationAndWeighting == null)
    return null;
  LCIAMethodType type = mav.normalisationAndWeighting.type;
  return type == null ? null : type.value();
}
```