better move the regionalized flow index to the olca-geo module
we will develop the new regionalized calculation in this module as
an extension to the current calculation. however, later we may want
to merge this with the normal calculation so that the regionalized
version would just differ in the flow index and would be supported in
all different calculation types.