read and write locations of exchanges and LCIA
characterization factors in the ILCD import and
export.

we could remove our ILCD extensions. we have our JSON based
format and we have to make sure that it provides lossless
data exchange. thus, we do not have to try this for the ILCD
format -> just read and write to and from this format as
good as we can.