this does not really work for project results:
```java
@Override
public List<IndexFlow> getFlows() {
  ...
}
```

as each contribution has its own index order
this can lead to quite some problems...

how to solve this?

maybe like this: for the results with a single
flow index we can directly use the `index` field
to locate values in matrices and vectors

a project result is a multi-flow index result.
there we use the flow and location descriptors
to locate the values in the respective sub-results.