# Contributions by Locations
In a regionalized calculation we map `(flow, location)` pairs to matrix rows and
columns instead of just flows as in the non-regionalized calculation. The
matrices in the results are then the same but the calculation of location
contributions are different.

## Location contributions of a flow result
The _direct flow results_ can be calculated via:

$$\mathbf{G} = \mathbf{B} \ \text{diag}(\mathbf{s})$$

The matrix has the following shape:

```
             products
            +---------+
            |         |
elem. flows |         |
            |         |
            +---------+
```

In the non-regionalized calculation each flow is mapped to a single row $i$.
The total result $g_i$ of this flow $i$ is the sum of the corresponding row in
$G$:

$$g_i = \sum_j G(i,j)$$

To calculate the location contributions we take the locations of the
processes of the products in the column index of $G$:

```
             products
            +---------+
            |         |
elem. flows |||||||||||
            |         |
            +---------+
             | || | |
             +--+---+--> location k
               +--+----> location l
```

In the regionalized calculation, flow-location pairs are mapped to the row
index of $G$. The total result of a flow is then the sum of multiple rows and
the contribution of a location to that flow is just the sum of the corresponding
row:

```
             products
            +---------+
            |         |
elem. flows |---------| -> location k
            |---------| -> location l
            +---------+
```