we can maybe simplify the export of references from parameter
redefinitions a bit. If these are global parameters => we need
to export them; if these a process parameters we do not need
to export the respective processes because they must be already
referenced by the respective product systems. if these are LCIA
methods, then we need to export them. currently, LCIA categories
can be only referenced by projects. as we may switch to scenarios
only for parameter redefinitions, we may do not have redefinitions
of LCIA category parameters in the future ... and this is not that
bad.