the problem is in DQData ...
also, we probably could save quite some memory when we store things in simple
byte arrays. the number of dq indicators is very small (max. ~ 10) and also the
number of scores. we just need to hold the exchange dq scores of elementary
flows and the process dq scores respectively (also: why do we need process dq
scores? couldn't this just be the dq scores of the reference flows?).

however, we basically just need the flow and tech. index of a result to collect
the scores of each indicator in a byte matrix: with the flow index mapped to the
rows and the tech. index mapped to the columns. the we aggregate thes matrices
based on the results. that should be quite easy, much faster, and using much
less memory as the current implementation.