added some compatibility methods
in the long term it would be good to have the result views based on
`ProcessProducts` and not just on processes. For mono-funtional
databases this is fine, but otherwise there can be multiple instances of
a process in the result: one for each product; and it would be good to
see this in the result. this is also related to libaries, where we build
the tech.-index also based on `ProcessProduct` instances