this needs tests but the idea is to expand the
significant upstream paths with only a minimal
amount of calculations. when there is a limit
regarding the max. number of nodes sets, it uses
a priority queue to find the best paths that
should be expanded. it will be interesting to
see how this performs with the lazy solution
provider.