when we access columns of sparse matrices,
we load the complete thing into memory currently. this is correct but we should
also cache it for the next read. we could maintain a simple cache for this in
the `Library` class.

```java
private final Map<LibraryMatrix, IMatrix> matrixCache = new HashMap<>();

private IMatrix getMatrix(LibraryMatrix name) {
  var m = matrixCache.get(name);
  if (m != null)
    return m;
  // ...

  matrixCache.put(name, m);
  return m;
}
```

another option would be to have something like a `MatrixFile` that would eagerly
load sparse matrices into memory. for dense matrices it could load the thing
into memory when it is small or when there are a lot of reads on it. but for
now we may go with the simple matrix cache.