we will replace the data in the results with function calls
in the `SolutionProvider`. in the old results it was assumed that all data is
present. but now things can be calculated or loaded from disk on demand. a
result will just wrap a corresponding `SolutionProvider`. we will try to keep
most of the result methods in the results stable and delegate them to the
solution provider...