a file based block matrix for $B$
the flow index $\Phi$ of the combined system defines the rows of the combined
intervention matrix $B$; the tech. index $\Tau$ the columns. the flow index of
the foreground matrix $\Phi_F$ is the first part of $\Phi$, the rest is the flow index
of the background system $\Phi_B$ (which can be a combination of several libraries)
without the flows that are already part of $\Phi_F$ :

$$
\Phi = \begin{bmatrix} \Phi_F \\ \Phi_B \backslash \Phi_F \end{bmatrix}
$$

the same is true for the tech. index:

$$
\Tau = \begin{bmatrix} \Tau_F \\ \Tau_B \backslash \Tau_F \end{bmatrix}
$$

the column $b_j$ of a process in the foreground system is then:

$$
b_j = \begin{bmatrix} b_{jF} \\ 0 \end{bmatrix}
$$

if $j$ is in the background system we have to map the indices for the flow $f$:

$$
b_j[\Phi[f]] = b_{jB}[\Phi_B[f]]
$$

the same needs to be done to get the column index $j_B$ of the product $p$ in
the background system:

$$
j_B = \Tau_B[p]
$$

we just ask the respective library for the column $b_{jB}$. currently, it loads
sparse matrices completely into memory and caches them, while it seeks into
the files for dense matrices... we could maybe also cache smaller dense
matrices in memory to avoid disk reads but this is a library internal thing...
but the thing is that we already have the file based block matrix structure
with these mechanisms.