so... we can get an ei3.7 LCI result in __153 ms__
with the library calculation. a nice next step would be to test this with
the new `Sankey` calculation. for this we just need to seek into $M_B$ for the
background processes...

it is basically how we calculate the total flow result $g$ but scaled to the
solution $s_j = A^{-1}[:,j]$. calculating $g$ is just a special case of this as we
just need to scale $m_j$ with the final demand $f_j$

$$
g = m_j \odot f_j
$$

yes, we get the same result:

```java
var result = provider.totalFlowResults();
var ofOne = provider.totalFlowResultsOfOne(0);
for (int i = 0; i < result.length; i++) {
  var epsilon = Math.abs(ofOne[i] - result[i]);
  if (epsilon > 1e12)
    System.out.println("eps > 1e12");
    System.exit(1);
  }
}
```