now this error:
```
Exception in thread "grpc-default-executor-0" java.lang.NoSuchMethodError:
 'boolean org.openlca.proto.generated.Proto$ProductSystem.access$8()'
  at org.openlca.proto.generated.Proto$ProductSystem$Builder.maybeForceBuilderInitialization(Proto.java:18523)
  at org.openlca.proto.generated.Proto$ProductSystem$Builder.<init>(Proto.java:18513)
```

this is this? https://stackoverflow.com/q/59880287

when running `mvn package` again it works. but does it land correctly in the
app? seems so. do not know what the error was here.

copy a simple system:

```py
import olcarpc as rpc
import uuid

if __name__ == '__main__':
    client = rpc.Client()
    sys: rpc.ProductSystem = client.get_product_system(name='p1').product_system
    sys.id = str(uuid.uuid4())
    sys.name = "copied thing"
    client.put_product_system(sys)
    print(sys.name)
    client.close()
```

works, but we only have a basic import for now.

but I get this in the graphical editor:

```
!ENTRY org.eclipse.e4.ui.workbench 4 2 2021-01-14 09:49:53.431
!MESSAGE Problems occurred when invoking code from plug-in: "org.eclipse.e4.ui.workbench".
!STACK 0
java.lang.NullPointerException: Cannot invoke
  "org.eclipse.ui.IWorkbenchWindow.getSelectionService()" because the
  return value of "org.eclipse.ui.IWorkbenchPartSite.getWorkbenchWindow()" is null
  at org.eclipse.gef.ui.actions.SelectionAction.update(SelectionAction.java:125)
  at org.eclipse.gef.ui.parts.GraphicalEditor.updateActions(GraphicalEditor.java:458)
```