go back to a single `Ref` type
but with type specific fields (which are just empty in other types). inheritance is great but does not work with Protocol Buffers. so this is a simple pragmatic solution for exchanging data over gRPC (e.g. in a search we want to always get the type specific fields but the return type of a service should be just `stream Ref`).