adding `db.clearCache()` calls directly in
the `NativeSql` methods. when we modify a database via plain SQL statements we
always should clear the cache (at least when we are in the UI mode). otherwise
this can have strange effects when outdated objects are returned by the entity
manager. clearing the cache after these calls should be safe... let's see.