Introduce MatrixReader
The user of a result function should never modify the
returned data because we may just return the underlying
result matrix. With `MatrixReader` we can just enable
the functions for reading...

Maybe it would be good to also have VectorReader and
Vector?

but then there is the question: what should copy return?
co-variance!

```
@Override
Matrix copy();
```