we can forward the index calls to the providers
so that we do not need to store them in additional fields. we may want
to do the same with the scaling vector etc. also, there is no need for
the project results to implement the result interface, it is not used.
we also should check if the methods in the BaseResult are used at all...