rename store to reader
we split the `ProtoStore` interface into a `ProtoReader` and `ProtoWriter`
part. these parts are really independent and we always know when to use
what. There can be implementations like the `InMemoryProtoStore` that implement
both things but e.g. for the ZipStore it could be better and more efficient
to split this into a `ZipProtoReader` and `ZipProtoWriter` e.g. not using
the `FileStore` implementation anymore which has some problems...