remove the demand logic from the TechIndex
we need the possibility to freely set the demand to any position of the
technology index. we want this, so that we can point into library blocks
without re-ordering them when the reference process of a system is just
a library process. also, we want to align library blocks in any order
when we want to combine them efficiently.