spcsv: use the unit instead of the quantity in the mapping key
Reason is that the same flow can be given in different units, and
then we need different conversion factors in the mapping. Now, the
target flow could be given in another quantity and, thus, it is
not that easy to infer a conversion factor. While possible, the
mapping machinery currently has no knowledge of unit conversions,
so we just use the units directly in the mapping keys for now