remove SimpleResult usages so that we can merge it with `FullResult`
the switch for the calculation type is in the ResultProvider now;
this can be lazy (e.g. based on sparse factorization; lazy library
scaling) or eager (e.g. based on inversion and intensity calculations;
library block inversion).

we could store this type in the setup:

```
CalculationSetup.lazy(system);
```

but e.g. for Monte-Carlo-Simulations it is always lazy so this would
be ignored then; so better

```
var setup = CalculationSetup.of(system);
var result = SystemCalculator.lazy(setup, db).calculate();
var result = Simulator.of(setup, db).next();
```