added `DqiData`
an instance of this class will hold the raw data quality values of a product
system stored in byte matrices. we need to separate these matrices from the
current `DqResult` because we also want to write them in the matrix / library
exports. We may attach them to MatrixData so that it is easy to export them.