extracted more static modules as separate projects
in the future we want to push versions to Maven more regularly but
it then makes not sense to push things that did not change with a
new version. thus, we will extract the modules into separate
projects starting with the ones that change rarely. we also need a
better version management then.