ipc-parameter context: default to process when `@type` is missing
currently, the olca-ipc.ts API does not accept parameter definitions with
a defined context when the `@type` attribute is missing. With this fix we
default to process parameters in this case as this is the most common use
case. However, we should also fix this in the olca-ipc.ts package as it
currently does not seem to deserialize the `@type` attributes.