first experimental draft of a ProviderChainRemoval
the idea is as follows: if we remove a link, we build a cluster for the provider `P` in the link: recursively add all processes that link to `P` or are linked from `P`. If the reference process `R` of the system is not contained in the cluster, we can remove the cluster from the graph. Otherwise, when we find `R` during the build of the cluster, we directly stop and do nothing. This is in principle what is already done, but with smarter traversal and indexing it could be done faster.

It currently uses collection types from the SDK but we could move to primitive collections if they are significantly faster