new ProcessDoc structure
This is the first draft of a new structure for process metadata
fields. At some point we may switch to a general ModelDoc class
that could be also used in EPDs (and product systems) for some
common fields. The current changes are backwards compatible so
that we do not have to run updates on the openLCA JSON schema.
We may break that compatibility in a future version (e.g. to
support multiple reviews per dataset). Other metadata fields
may be written to extension fields, this is why this includes
the =propext= branch. This update requires the following
database changes:

+ rename/copy column: =completeness -> data_completeness=
+ rename/copy column: =sampling -> sampling_procedure=
+ add column: =useAdvice=
+ add column: =review_type=
+ add column: =f_review_report=
+ add table: =tbl_compliance_declarations=
+ rename column: =f_dataset_owner -> f_data_owner=

Also, the new fields need to be added to the openLCA schema and
the various imports and exports.