bug fix in Sankey
When the number of nodes is limited we use a priority
queue for nodes that can be added next when expanding
the graph. In this queue, there can be duplicates of
provider nodes to be added next under some parent nodes.
When adding such a provider node, we must only increase
the number of nodes in the graph, if that provider was
not added before. This was not correctly done before.