rethink the number generator for the log-normal distribution
Lets reduce the mean value of the underlying normal
distribution by sigma^2/2. As the mean value of the
log-normal distribution is exp(µ + sigma^2/2) we then
get a mean value of the resulting distribution that is
the same as the median (which is the geometric mean,
entered by the user). This makes much more sense in
the results of the Monte Carlo Simulation then.

The same principle is applied in the
`reparametrize_lognormal_to_agree_with_static_amount`
function of the EcoSpold 2 import in Brightway:

https://github.com/brightway-lca/brightway2-io/blob/a3f5db3ba29e7b7f1a41b535605fd4f8e11a20d4/bw2io/strategies/ecospold2.py#L860